requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
pandas==2.0.3
openpyxl==3.1.2
Email-Validator==2.0.0
Pillow==10.0.0